requirements.txt
openpyxl>=3.1.0
ezdxf>=0.18.0
pdf2image>=1.16.0
Pillow>=10.0.0
openai>=1.0.0
python-dotenv>=1.0.0
